Add files via upload
The dataset consists of several predictor variables and one target variable, Diagnosis. 
The target variable has values 'Benign' and 'Malignant', where 'Benign' means that the 
cells are not harmful or there is no cancer and 'Malignant' means that the patient has 
cancer and the cells have a harmful effect. 
Given the details of cell nuclei taken from breast mass, predicted whether or not a patient 
has breast cancer using the Ensemble Techniques and evaluated the model based on performance 
metrics other than model accuracy.